Improve Sub Handling
Fiddle with sub movement settings to hopefully achieve a better result. Sub moves and turns faster, as well as a slightly better pitch/anti-roll system.
Added a switch sound when flipping a switch, and also added a ping sound for the start of the sonar system.
The player can now flip switches and drag levers in-game (and the reactor starts on by default).
Fix up mouse movement a bit so looping is handled correctly, and add some arrows when interacting with levers
Fix light switches being stuck if power is off.
Added a makeshift compass and depth gauge as placeholders until I get around to re-making the helm to implement them. The new helm should also have a pitch indicator, as well as a speedometer of some kind.
Probably some other stuff that I'm missing as well, who knows.